requirements.txt
pillow~=10.4.0
pip~=23.1.2